+++ src/depth_buffer.h
@@ -4,13 +4,13 @@
 
 typedef struct
 {
-    uint16_t width;
-    uint16_t height;
+    int width;
+    int height;
     float* data;
 
 } depthbuffer;
 
-depthbuffer* depthbuffer_new(uint16_t width, uint16_t height);
+depthbuffer* depthbuffer_new(int width, int height);
 void depthbuffer_set(depthbuffer* buffer, int x, int y, float val);
 float depthbuffer_get(depthbuffer* buffer, int x, int y);
 void depthbuffer_free(depthbuffer* buffer);
